# ketroy_back_new/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API документация Ketroy",
        "description": "Документация API для управления администраторами.",
        "version": "1.0"
    },
    "paths": {
...
        "/reviews": {
            "get": {
                "tags": [
                    "ShopReviews"
                ],
                "summary": "Получение всех отзывов о магазинах",
                "description": "Возвращает список всех отзывов о магазинах.",
                "operationId": "912c8a812e1c72243382ea458b80fbb7",
                "responses": {
                    "200": {
                        "description": "Список отзывов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "shop_id": {
                                                "type": "integer"
                                            },
                                            "user_id": {
                                                "type": "integer"
                                            },
                                            "review": {
                                                "type": "string"
                                            },
                                            "shop": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    },
                                                    "city": {
                                                        "type": "string"
                                                    },
                                                    "type": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "user": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "ShopReviews"
                ],
                "summary": "Добавление отзыва о магазине",
                "description": "Создает новый отзыв о магазине.",
                "operationId": "eb0f98143b2f487adea14c9ef2c9f9b2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "shop_id",
                                    "user_id",
                                    "review"
                                ],
                                "properties": {
                                    "shop_id": {
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "review": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Отзыв успешно добавлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Отзыв успешно добавлен!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/reviews/{id}": {
            "delete": {
                "tags": [
                    "ShopReviews"
                ],
                "summary": "Удаление отзыва",
                "description": "Удаляет отзыв по указанному идентификатору.",
                "operationId": "53c9ca5ade108f0e2bbcb2b8c01bc364",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID отзыва",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Отзыв успешно удалён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Отзыв успешно удалён"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Отзыв не найден"
                    }
                }
            }
        },
...
    },
...
}